refactor(c-api): align kth_wallet_ec_to_address version with C++ (#415)
The C-API signature took `uint32_t version` for what the C++ wrapper
(`payment_address::from_ec_public(ec_public, uint8_t)`) expects as a
single-byte address prefix. The extra width forced a
`static_cast<uint8_t>` inside the shim.

Change the C-API signature to `uint8_t version` so the types line up.
The cast disappears. ABI break for consumers is intentional — the
C-API mirrors the C++ shape; there are no shim widenings to be
preserved.

No behavior change; kth_capi_test 769/769 pass.

+++ src/c-api/include/kth/capi/wallet/wallet.h
@@ -39,7 +39,7 @@ KTH_EXPORT
 kth_ec_public_t kth_wallet_ec_to_public(kth_ec_secret_t secret, kth_bool_t uncompressed);
 
 KTH_EXPORT
-kth_payment_address_t kth_wallet_ec_to_address(kth_ec_public_t point, uint32_t version);
+kth_payment_address_t kth_wallet_ec_to_address(kth_ec_public_t point, uint8_t version);
 
 
 // KTH_EXPORT
